[BEAM-8666] Remove dependency between DataflowRunner and PortableRunner introduced by PR#9811
PR#9811 creates a new dependency between `DataflowRunner` and `PortableRunner`. `DataflowRunner` uses `DockerEnvironment` but there was no need for `DataflowRunner` to depend on the `PortableRunner` module before. We will refactor out `default_docker_image()` method to live in `apache_beam.transforms.Environment`.
